style: move function
