+++ src/decorator/number/IsSafeInt.ts
